// src/app/modules/auth/interfaces/UserInfo.ts
export interface UserInfo {
    email: string;
    password: string | number;
    nickname: string;
    first_name: string;
    last_name: string;
    phone: number;
    gender_orientation: string;
    city: string;
    country: string;
    date_of_birth_day: number;
    date_of_birth_month: number;
    date_of_birth_year: number;
}